requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
builtwith==1.3.4
certifi==2024.2.2
charset-normalizer==3.3.2
dbus-python==1.3.2
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
plyer==2.1.0
pygame==2.5.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-whois==0.9.4
requests==2.31.0
rich==13.7.1
six==1.16.0
soupsieve==2.5
urllib3==2.2.1
